move EvolveOptions to methods and update it's dependencies
