+++ src/main/java/com/github/raimannma/methods/EvolveOptions.java
@@ -1,11 +1,9 @@
-package com.github.raimannma.architecture;
+package com.github.raimannma.methods;
 
 import java.util.Arrays;
 import java.util.Objects;
 import java.util.function.ToDoubleFunction;
-import com.github.raimannma.methods.Loss;
-import com.github.raimannma.methods.Mutation;
-import com.github.raimannma.methods.Selection;
+import com.github.raimannma.architecture.Network;
 
 /**
  * The type Evolve options.
