+++ src/test/java/com/github/raimannma/architecture/NEATTest.java
@@ -5,6 +5,7 @@ import static com.github.raimannma.methods.Utils.setsEqual;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.stream.IntStream;
+import com.github.raimannma.methods.EvolveOptions;
 import com.github.raimannma.methods.Mutation;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
